put map decompressing and parsing in the cache mod
